@@ -4,46 +4,49 @@
 
 import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
-import 'package:openflutterecommerce/data/abstract/product_repository.dart';
+import 'package:openflutterecommerce/data/abstract/favorites_repository.dart';
+import 'package:openflutterecommerce/data/abstract/model/sort_rules.dart';
 import 'package:openflutterecommerce/data/fake_model/hashtag_repository.dart';
 
 import 'favorites_event.dart';
 import 'favorites_state.dart';
 
 class FavouriteBloc extends Bloc<FavouriteEvent, FavouriteState> {
-  final ProductRepository favouriteRepository;
+  final FavoritesRepository favoritesRepository;
   final HashtagRepository hashtagRepository;
 
   FavouriteBloc({
-    @required this.favouriteRepository,
+    @required this.favoritesRepository,
     @required this.hashtagRepository,
-  }) : assert(favouriteRepository != null);
+  });
+
+  @override
+  FavouriteState get initialState => FavouriteState();
 
   @override
   Stream<FavouriteState> mapEventToState(FavouriteEvent event) async* {
-    print('mapEventToState: $event');
-    if (event is ShowListViewEvent) {
-      yield FavouriteListViewState(await favouriteRepository.getFavorites(),
-          hashtagRepository.getHashtags());
-    } else if (event is ShowTileViewEvent) {
-      yield FavouriteTileViewState(await favouriteRepository.getFavorites(),
-          hashtagRepository.getHashtags());
-    } else if (event is MakeFavoriteEvent) {
-      if (event.isFavorite) {
-        await favouriteRepository.addToFavorites(event.productId);
-      } else {
-        await favouriteRepository.removeFromFavorites(event.productId);
-      }
-      if (state is FavouriteListViewState) {
-        yield FavouriteListViewState(await favouriteRepository.getFavorites(),
-            hashtagRepository.getHashtags());
-      } else {
-        yield FavouriteTileViewState(await favouriteRepository.getFavorites(),
-            hashtagRepository.getHashtags());
-      }
+    if (event is ScreenLoadedEvent) {
+      yield FavouriteState(
+          sortBy: SortRules(),
+          data: await favoritesRepository.getFavoriteProducts(),
+          filterRules: await favoritesRepository.getFavoritesFilterOptions());
+    } else if (event is ProductsChangeViewEvent) {
+      yield state.copyWith(isList: !state.isList);
+    } else if (event is ProductChangeSortRulesEvent) {
+      yield state.getLoading();
+      final filteredData = await favoritesRepository.getFavoriteProducts(
+          filterRules: state.filterRules, sortRules: event.sortBy);
+      yield state.copyWith(
+        sortBy: event.sortBy,
+        data: filteredData,
+      );
+    } else if (event is ProductChangeFilterRulesEvent) {
+      yield state.getLoading();
+      final filteredData = await favoritesRepository.getFavoriteProducts(
+          filterRules: event.filterRules, sortRules: state.sortBy);
+      yield state.copyWith(filterRules: event.filterRules, data: filteredData);
+    } else if (event is AddToCartEvent) {
+      //TODO add to cart
     }
   }
-
-  @override
-  FavouriteState get initialState => InitialViewState();
 }
