@@ -5,21 +5,24 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_translate/flutter_translate.dart';
 import 'package:openflutterecommerce/config/routes.dart';
 import 'package:openflutterecommerce/config/theme.dart';
+import 'package:openflutterecommerce/data/abstract/favorites_repository.dart';
+import 'package:openflutterecommerce/data/abstract/model/filter_rules.dart';
 import 'package:openflutterecommerce/data/abstract/product_repository.dart';
 import 'package:openflutterecommerce/data/abstract/user_repository.dart';
 import 'package:openflutterecommerce/data/fake_model/fake_product_repository.dart';
 import 'package:openflutterecommerce/data/fake_model/fake_user_repository.dart';
 import 'package:openflutterecommerce/features/sign_in/sign_in.dart';
+import 'package:openflutterecommerce/presentation/features/filters/filters_screen.dart';
 import 'package:openflutterecommerce/presentation/features/product_details/product_screen.dart';
 import 'package:openflutterecommerce/presentation/features/products/products.dart';
 import 'package:openflutterecommerce/presentation/features/splash_screen.dart';
 
-import 'data/abstract/cart_repository.dart';
-import 'data/fake_model/fake_cart_repository.dart';
-import 'features/authentication/authentication.dart';
 import 'config/routes.dart';
+import 'data/abstract/cart_repository.dart';
 import 'data/abstract/category_repository.dart';
+import 'data/fake_model/fake_cart_repository.dart';
 import 'data/fake_model/fake_category_repository.dart';
+import 'features/authentication/authentication.dart';
 import 'features/forget_password/forget_password.dart';
 import 'features/sign_up/sign_up_bloc.dart';
 import 'presentation/features/cart/cart.dart';
@@ -69,6 +72,9 @@ void main() async {
             create: (context) => FakeCategoryRepository(),
           ),
           RepositoryProvider<ProductRepository>(
+            create: (context) => FakeProductRepository(),
+          ),
+          RepositoryProvider<FavoritesRepository>(
             create: (context) => FakeProductRepository(),
           ),
           RepositoryProvider<UserRepository>(
@@ -185,6 +191,11 @@ class OpenFlutterEcommerceApp extends StatelessWidget {
       final ProductDetailsParameters parameters = settings.arguments;
       return MaterialPageRoute(builder: (context) {
         return ProductDetailsScreen(parameters);
+      });
+    } else if (settings.name == OpenFlutterEcommerceRoutes.filters) {
+      final FilterRules filterRules = settings.arguments;
+      return MaterialPageRoute(builder: (context) {
+        return FiltersScreen(filterRules);
       });
     } else {
       return MaterialPageRoute(
