@@ -4,7 +4,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:openflutterecommerce/data/abstract/product_repository.dart';
+import 'package:openflutterecommerce/data/abstract/cart_repository.dart';
 import 'package:openflutterecommerce/presentation/features/wrapper.dart';
 import 'package:openflutterecommerce/presentation/widgets/widgets.dart';
 
@@ -25,9 +25,7 @@ class _CheckoutScreenState extends State<CheckoutScreen> {
       title: 'Checkout',
       body: BlocProvider<CheckoutBloc>(
           create: (context) {
-            return CheckoutBloc(
-                productRepository:
-                    RepositoryProvider.of<ProductRepository>(context))
+            return CheckoutBloc(RepositoryProvider.of<CartRepository>(context))
               ..add(CheckoutStartEvent());
           },
           child: CheckoutWrapper()),
