@@ -4,41 +4,52 @@
 
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
+import 'package:openflutterecommerce/data/abstract/model/favorite_product.dart';
+import 'package:openflutterecommerce/data/abstract/model/filter_rules.dart';
 import 'package:openflutterecommerce/data/abstract/model/hashtag.dart';
-import 'package:openflutterecommerce/data/abstract/model/product.dart';
+import 'package:openflutterecommerce/data/abstract/model/sort_rules.dart';
 
 @immutable
-abstract class FavouriteState extends Equatable {
-  final List<Product> favouriteProducts;
+class FavouriteState extends Equatable {
+  final List<FavoriteProduct> data;
   final List<HashTag> hashtags;
+  final SortRules sortBy;
+  final FilterRules filterRules;
+  final String error;
+  final bool isList;
 
-  FavouriteState(this.favouriteProducts, this.hashtags);
+  FavouriteState(
+      {this.data,
+      this.hashtags,
+      this.sortBy,
+      this.filterRules,
+      this.error,
+      this.isList = true});
+
+  FavouriteState copyWith({
+    List<FavoriteProduct> data,
+    List<HashTag> hashtags,
+    SortRules sortBy,
+    FilterRules filterRules,
+    String error,
+    bool isList,
+  }) {
+    return FavouriteState(
+        data: data ?? this.data,
+        hashtags: hashtags ?? this.hashtags,
+        sortBy: sortBy ?? this.sortBy,
+        filterRules: filterRules ?? this.filterRules,
+        isList: isList ?? this.isList);
+  }
+
+  bool get isProductsLoading => data == null;
+  bool get isFilterRulesVisible => filterRules != null;
+  bool get hasError => error != null;
+
+  FavouriteState getLoading() {
+    return copyWith(data: null);
+  }
 
   @override
-  List<Object> get props => [favouriteProducts, hashtags];
-}
-
-@immutable
-class InitialViewState extends FavouriteState {
-  InitialViewState() : super(null, null);
-}
-
-@immutable
-class FavouriteListViewState extends FavouriteState {
-  FavouriteListViewState(
-      List<Product> favouriteProducts, List<HashTag> hashtags)
-      : super(favouriteProducts, hashtags);
-
-  @override
-  String toString() => 'FavouriteListViewState';
-}
-
-@immutable
-class FavouriteTileViewState extends FavouriteState {
-  FavouriteTileViewState(
-      List<Product> favouriteProducts, List<HashTag> hashtags)
-      : super(favouriteProducts, hashtags);
-
-  @override
-  String toString() => 'FavouriteTileViewState';
+  List<Object> get props => [data, hashtags, sortBy, filterRules, error];
 }
