@@ -23,8 +23,8 @@ class HomeBloc extends Bloc<HomeEvent, HomeState> {
     if (event is HomeLoadEvent) {
       if (state is HomeInitialState) {
         yield HomeLoadedState(
-            salesProducts: await productRepository.getProductsInCategory(1),
-            newProducts: await productRepository.getProductsInCategory(2));
+            salesProducts: await productRepository.getProducts(categoryId: 1),
+            newProducts: await productRepository.getProducts(categoryId: 2));
       } else if (state is HomeLoadedState) {
         yield state;
       }
