@@ -1,4 +1,5 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:openflutterecommerce/data/abstract/favorites_repository.dart';
 import 'package:openflutterecommerce/data/abstract/product_repository.dart';
 
 import 'product_event.dart';
@@ -6,9 +7,11 @@ import 'product_state.dart';
 
 class ProductBloc extends Bloc<ProductEvent, ProductState> {
   final ProductRepository productRepository;
+  final FavoritesRepository favoriesRepository;
   final int productId;
 
-  ProductBloc({this.productId, this.productRepository});
+  ProductBloc(
+      {this.favoriesRepository, this.productId, this.productRepository});
 
   @override
   ProductState get initialState => ProductInitialState();
@@ -22,9 +25,10 @@ class ProductBloc extends Bloc<ProductEvent, ProductState> {
           await productRepository.getSimilarProducts(productId);
       yield ProductLoadedState(product: data, similarProducts: similarProducts);
     } else if (event is ProductAddToFavoritesEvent) {
-      await productRepository.addToFavorites(productId);
+      await favoriesRepository.addToFavorites(
+          productId, event.selectedAttributes);
     } else if (event is ProductAddToFavoritesEvent) {
-      await productRepository.removeFromFavorites(productId);
+      await favoriesRepository.removeFromFavorites(productId);
     } else if (event is ProductAddToCartEvent) {
       //TODO: add to cart
     } else if (event is ProductSetAttributeEvent) {
