@@ -3,4 +3,4 @@ export 'products_event.dart';
 export 'products_screen.dart';
 export 'products_state.dart';
 export 'views/list_view.dart';
-export 'views/card_view.dart';
+export 'views/tile_view.dart';
