@@ -5,14 +5,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:openflutterecommerce/data/abstract/category_repository.dart';
+import 'package:openflutterecommerce/data/abstract/favorites_repository.dart';
 import 'package:openflutterecommerce/data/abstract/product_repository.dart';
 import 'package:openflutterecommerce/data/fake_model/hashtag_repository.dart';
-import 'package:openflutterecommerce/presentation/features/wrapper.dart';
+import 'package:openflutterecommerce/presentation/widgets/data_driven/size_changing_app_bar.dart';
+import 'package:openflutterecommerce/presentation/widgets/independent/error_dialog.dart';
 import 'package:openflutterecommerce/presentation/widgets/widgets.dart';
 
 import 'products.dart';
-import 'views/brands.dart';
-import 'views/filters.dart';
 
 class ProductsScreen extends StatefulWidget {
   final ProductListScreenParameters parameters;
@@ -35,7 +35,7 @@ class _ProductsScreenState extends State<ProductsScreen> {
     return SafeArea(
         child: OpenFlutterScaffold(
       background: null,
-      title: 'Products',
+      title: null,
       body: BlocProvider<ProductsBloc>(
           create: (context) {
             return ProductsBloc(
@@ -44,34 +44,46 @@ class _ProductsScreenState extends State<ProductsScreen> {
                     RepositoryProvider.of<ProductRepository>(context),
                 categoryRepository:
                     RepositoryProvider.of<CategoryRepository>(context),
+                favoritesRepository:
+                    RepositoryProvider.of<FavoritesRepository>(context),
                 hashtagRepository: HashtagRepository())
               ..add(ScreenLoadedEvent());
           },
-          child: ProductsWrapper()),
+          child: BlocConsumer<ProductsBloc, ProductsState>(
+            listener: (context, state) {
+              if (state.hasError) {
+                ErrorDialog.showErrorDialog(context, state.error);
+              }
+            },
+            builder: (context, state) {
+              return CustomScrollView(
+                slivers: <Widget>[
+                  SizeChangingAppBar(
+                    title: state.data?.category?.name,
+                    filterRules: state.filterRules,
+                    sortRules: state.sortBy,
+                    isListView: state is ProductsListViewState,
+                    onFilterRulesChanged: (filter) {
+                      BlocProvider.of<ProductsBloc>(context)
+                          .add(ProductChangeFilterRulesEvent(filter));
+                    },
+                    onSortRulesChanged: (sort) {
+                      BlocProvider.of<ProductsBloc>(context)
+                          .add(ProductChangeSortRulesEvent(sort));
+                    },
+                    onViewChanged: () {
+                      BlocProvider.of<ProductsBloc>(context)
+                          .add(ProductsChangeViewEvent());
+                    },
+                  ),
+                  state is ProductsListViewState
+                      ? ProductsListView()
+                      : ProductsTileView(),
+                ],
+              );
+            },
+          )),
       bottomMenuIndex: 1,
     ));
-  }
-}
-
-class ProductsWrapper extends StatefulWidget {
-  @override
-  _ProductsWrapperState createState() => _ProductsWrapperState();
-}
-
-class _ProductsWrapperState extends OpenFlutterWrapperState<ProductsWrapper> {
-  //State createState() => OpenFlutterWrapperState();
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<ProductsBloc, ProductState>(
-        bloc: BlocProvider.of<ProductsBloc>(context),
-        builder: (BuildContext context, ProductState state) {
-          return getPageView(<Widget>[
-            ProductsListView(changeView: changePage),
-            ProductsCardView(changeView: changePage),
-            ProductFilterView(changeView: changePage),
-            SelectBrandView(changeView: changePage)
-          ]);
-        });
   }
 }
