@@ -3,16 +3,16 @@
 // Date: 2020-02-06
 
 import 'package:flutter/material.dart';
+import 'package:openflutterecommerce/config/routes.dart';
 import 'package:openflutterecommerce/config/theme.dart';
 import 'package:openflutterecommerce/data/abstract/model/product.dart';
+import 'package:openflutterecommerce/presentation/features/product_details/product_screen.dart';
 
 import '../widgets.dart';
 
 class OpenFlutterProductListView extends StatelessWidget {
   final double width;
   final double height = 284;
-  final double elementHeight = 184;
-  final double elementWidth = 148;
   final List<Product> products;
 
   const OpenFlutterProductListView({
@@ -23,18 +23,23 @@ class OpenFlutterProductListView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    var productTiles = <Widget>[];
-    for (var i = 0; i < products.length; i++) {
-      productTiles.add(
-        OpenFlutterProductCard(
-            width: elementWidth, height: elementHeight, product: products[i]),
-      );
-    }
+    print('product number is ${products.length}');
     return Container(
         padding: EdgeInsets.only(top: AppSizes.sidePadding),
         width: width,
         height: height,
-        child:
-            ListView(scrollDirection: Axis.horizontal, children: productTiles));
+        child: ListView(
+            scrollDirection: Axis.horizontal,
+            children: products
+                .map((product) => GeneralProductTile(
+                      product: product,
+                      onFavoritesClick: null, //TODO add favorites actions
+                      onClick: () {
+                        Navigator.of(context).pushNamed(
+                            OpenFlutterEcommerceRoutes.product,
+                            arguments: ProductDetailsParameters(product.id));
+                      },
+                    ))
+                .toList(growable: false)));
   }
 }
