@@ -24,37 +24,40 @@ class OpenFlutterBlockSubtitle extends StatelessWidget {
     var _theme = Theme.of(context);
     const rightLinkWidth = 100.0;
     return Container(
-        padding: EdgeInsets.only(
-            top: AppSizes.sidePadding, left: AppSizes.sidePadding),
+      padding: EdgeInsets.only(
+          top: AppSizes.sidePadding, left: AppSizes.sidePadding),
+      child: SingleChildScrollView(
         child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              Row(
-                children: <Widget>[
-                  Container(
-                      width: width - rightLinkWidth,
-                      child: Text(title,
-                          style: _theme.textTheme.display1.copyWith(
-                              fontWeight: FontWeight.bold,
-                              color: _theme.primaryColor))),
-                  linkText != null
-                      ? InkWell(
-                          onTap: (() => {onLinkTap()}),
-                          child: Container(
-                            width: rightLinkWidth,
-                            child: Align(
-                              alignment: Alignment.centerRight,
-                              child: Text(
-                                linkText,
-                                style: _theme.textTheme.display1
-                                    .copyWith(color: _theme.accentColor),
-                              ),
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            Row(
+              children: <Widget>[
+                Container(
+                    child: Text(title,
+                        style: _theme.textTheme.display1.copyWith(
+                            fontWeight: FontWeight.bold,
+                            color: _theme.primaryColor))),
+                linkText != null
+                    ? InkWell(
+                        onTap: (() => {onLinkTap()}),
+                        child: Container(
+                          width: rightLinkWidth,
+                          child: Align(
+                            alignment: Alignment.centerRight,
+                            child: Text(
+                              linkText,
+                              style: _theme.textTheme.display1
+                                  .copyWith(color: _theme.accentColor),
                             ),
                           ),
-                        )
-                      : Container()
-                ],
-              )
-            ]));
+                        ),
+                      )
+                    : Container()
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
   }
 }
